+++ lib/helpers.js
@@ -1,208 +1,18 @@
 'use strict';
 
 const logger = require('heroku-logger');
-const Promise = require('bluebird');
-const request = require('request-promise');
 
 const contentful = require('./contentful');
 const gambitCampaigns = require('./gambit-campaigns');
-
 const config = require('../config/lib/helpers');
+const helpers = require('./helpers/index');
 
-/**
- * Attachments helper
- * // TODO: Should this live inside the exports.request object namespace?
- */
-module.exports.attachments = {
+// TODO: Move contents of this helper.js file into lib/helpers/index.js or to modular helpers
 
-  /**
-   * add - Adds attachments to its respective direction array.
-   *
-   * @param  {object} req
-   * @param  {object} attachmentObject object containing url, and contentType
-   * @param  {string} direction = 'inbound' direction array where to store this attachment
-   */
-  add: function add(req, attachmentObject, direction = 'inbound') {
-    if (attachmentObject.url) {
-      // TODO: req.attachments might not exist if this function is called from a route
-      // that doesnt set a default req.attachments
-      req.attachments[direction].push({
-        url: attachmentObject.url,
-        contentType: attachmentObject.contentType,
-      });
-    }
-  },
-
-  /**
-   * parseFromReq - creates a generic attachment object from a req which contains a mediaUrl in the
-   * body of the request.
-   *
-   * @param  {type} req description
-   * @return {type}     description
-   */
-  parseFromReq: function parseFromReq(req) {
-    return {
-      url: req.body.mediaUrl,
-    };
-  },
-};
-
-/**
- * Slack helper
- */
-module.exports.slack = {
-  /**
-   * parseBody - parses properties out of the body of a Slack request
-   *
-   * @param  {object} req
-   * @return {promise}
-   */
-  parseBody: function parseBody(req) {
-    req.platform = 'slack';
-    req.platformUserId = req.body.slackId;
-    req.slackChannel = req.body.slackChannel;
-    req.platformMessageId = req.body.messageId;
-    req.mediaUrl = req.body.mediaUrl;
-  },
-};
-
-/**
- * Facebook helper
- */
-module.exports.facebook = {
-  /**
-   * parseBody - parses properties out of the body of a Facebook request
-   *
-   * @param  {object} req
-   * @return {promise}
-   */
-  parseBody: function parseBody(req) {
-    req.platform = 'facebook';
-    req.platformUserId = req.body.facebookId;
-    req.platformMessageId = req.body.messageId;
-  },
-};
-
-/**
- * Customer.io helper
- */
-module.exports.customerIo = {
-  /**
-   * parseBody - parses properties out of the body of a Facebook request
-   *
-   * @param  {object} req
-   * @return {promise}
-   */
-  parseBody: function parseBody(req) {
-    req.platform = 'sms';
-    req.platformUserId = req.body.phone;
-    req.broadcastId = req.body.broadcastId;
-    req.messageFields = req.body.fields;
-  },
-};
-
-/**
- * Twilio helper
- */
-module.exports.twilio = {
-
-  /**
-   * parseBody - parses properties out of the body of a Twilio request
-   *
-   * @param  {object} req
-   * @return {promise}
-   */
-  parseBody: function parseBody(req) {
-    req.platform = 'sms';
-    req.platformUserId = req.body.From;
-    req.inboundMessageText = req.body.Body;
-    req.platformMessageId = req.body.MessageSid;
-
-    const attachmentObject = this.parseAttachmentFromReq(req);
-
-    return new Promise((resolve, reject) => {
-      if (attachmentObject.redirectUrl) {
-        exports.twilio.getAttachmentUrl(attachmentObject.redirectUrl)
-          .then((url) => {
-            req.mediaUrl = url;
-            attachmentObject.url = url;
-            exports.attachments.add(req, attachmentObject, 'inbound');
-            resolve(url);
-          })
-          .catch(error => reject(error));
-      } else {
-        resolve('attachmentObject does\'t contain a redirectUrl');
-      }
-    });
-  },
-
-  /**
-   * parseAttachmentFromReq - parses the MediaUrl0 and MediaContentType0 out of Twilio requests
-   * and creates an attachent object
-   *
-   * @param  {object} req
-   * @return {object}
-   */
-  parseAttachmentFromReq: function parseAttachmentFromReq(req) {
-    return {
-      redirectUrl: req.body.MediaUrl0,
-      contentType: req.body.MediaContentType0,
-    };
-  },
-
-  /**
-   * getAttachmentUrl - It makes a GET request to the redirectUrl uri. It redirects the request
-   * and returns the actual attachment's data. We are interested in storing the actual attachment's
-   * address, which we get from the full response object of the request.
-   *
-   * @param  {string} redirectUrl
-   * @return {promise}
-   */
-  getAttachmentUrl: function getAttachmentUrl(redirectUrl) {
-    const options = {
-      uri: redirectUrl,
-      // needed, otherwise returns the parsed body
-      resolveWithFullResponse: true,
-    };
-
-    return request(options)
-      .then((redirectRes) => {
-        let url = '';
-        try {
-          url = redirectRes.request.uri.href;
-        } catch (error) {
-          // The data structure might have changed and code needs to be updated.
-          logger.error('The Twilio attachment redirectUrl returns an error', error);
-        }
-        return url;
-      });
-  },
-};
-
-
-/**
- * Request helper
- */
-module.exports.request = {
-  isTwilio: function isTwilio(req) {
-    return !!req.body.MessageSid;
-  },
-  isSlack: function isSlack(req) {
-    return !!req.body.slackId;
-  },
-  isFacebook: function isFacebook(req) {
-    return !!req.body.facebookId;
-  },
-  isFront: function isFront(req) {
-    // TODO: Should be in a config constant
-    return !!req.get('x-front-signature');
-  },
-  isCustomerIo: function isCustomerIo(req) {
-    const platform = req.query.platform || req.body.platform;
-    // TODO: Should be in a config constant
-    return platform === 'customerio';
-  },
-};
+// register helpers
+Object.keys(helpers).forEach((helperName) => {
+  module.exports[helperName] = helpers[helperName];
+});
 
 module.exports.isAskContinueTemplate = function (templateName) {
   return config.askContinueTemplates.includes(templateName);
